REST vs GraphQL/task_manager_api/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.217.0
sqlalchemy==2.0.23
email-validator==2.1.0